v.0.0.3 'Added Controller Support'

+++ custom_types.py
@@ -5,6 +5,6 @@ from collections import defaultdict
 
 
 type ActionKeys = defaultdict[int, bool]
-type HoldKeys = defaultdict[int, float]
+type HoldKeys = defaultdict[int, int]
 
 type Coordinate = tuple[float, float] | p.Vector2
